@@ -13,17 +13,17 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^12.1.1",
-    "@angular/cdk": "^12.1.1",
-    "@angular/common": "^12.1.1",
-    "@angular/compiler": "^12.1.1",
-    "@angular/core": "^12.1.1",
+    "@angular/animations": "^12.2.11",
+    "@angular/cdk": "^12.2.11",
+    "@angular/common": "^12.2.11",
+    "@angular/compiler": "^12.2.11",
+    "@angular/core": "^12.2.11",
     "@angular/flex-layout": "^12.0.0-beta.34",
-    "@angular/forms": "^12.1.1",
-    "@angular/material": "^12.1.1",
-    "@angular/platform-browser": "^12.1.1",
-    "@angular/platform-browser-dynamic": "^12.1.1",
-    "@angular/router": "^12.1.1",
+    "@angular/forms": "^12.2.11",
+    "@angular/material": "^12.2.11",
+    "@angular/platform-browser": "^12.2.11",
+    "@angular/platform-browser-dynamic": "^12.2.11",
+    "@angular/router": "^12.2.11",
     "@ngx-translate/core": "^13.0.0",
     "@ngx-translate/http-loader": "^6.0.0",
     "chart.js": "^2.9.3",
@@ -33,7 +33,7 @@
     "domhandler": "^4.2.0",
     "domutils": "^2.1.0",
     "file-saver": "^2.0.2",
-    "htmlparser2": "^4.1.0",
+    "htmlparser2": "^7.1.2",
     "js-beautify": "^1.11.0",
     "lodash.clone": "^4.5.0",
     "lodash.foreach": "^4.5.0",
@@ -42,31 +42,31 @@
     "lodash.size": "^4.2.0",
     "lodash.slice": "^4.2.0",
     "lodash.without": "^4.4.0",
-    "ngx-gauge": "^3.0.0",
-    "rxjs": "^6.6.0",
+    "ngx-gauge": "^4.0.0",
+    "rxjs": "^7.4.0",
     "tslib": "^2.0.0",
     "web-animations-js": "^2.3.2",
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~12.1.1",
-    "@angular/cli": "~12.1.1",
-    "@angular/compiler-cli": "^12.1.1",
-    "@angular/language-service": "^12.1.1",
-    "@types/jasmine": "~3.6.0",
+    "@angular-devkit/build-angular": "~12.2.11",
+    "@angular/cli": "~12.2.11",
+    "@angular/compiler-cli": "^12.2.11",
+    "@angular/language-service": "^12.2.11",
+    "@types/jasmine": "^3.10.0",
     "@types/jasminewd2": "^2.0.8",
     "@types/js-beautify": "^1.11.0",
-    "@types/node": "^14.0.27",
+    "@types/node": "^16.11.3",
     "codelyzer": "^6.0.0",
-    "jasmine-core": "~3.6.0",
-    "jasmine-spec-reporter": "~5.0.0",
+    "jasmine-core": "^3.10.0",
+    "jasmine-spec-reporter": "^7.0.0",
     "karma": "^6.3.4",
     "karma-chrome-launcher": "~3.1.0",
     "karma-coverage-istanbul-reporter": "~3.0.2",
     "karma-jasmine": "~4.0.0",
     "karma-jasmine-html-reporter": "^1.5.0",
     "protractor": "~7.0.0",
-    "ts-node": "^8.10.2",
+    "ts-node": "^10.3.1",
     "tslint": "^6.1.2",
     "typescript": "~4.3.5"
   }
